+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased
+### Removed
+- PTR lookups no longer end with a dot '.'
+
+### Changed
+- Removed need for dig and replace with ruby resolv
 
 ## [0.0.4] - 2015-07-14
 ### Changed